Removed redundant casts

+++ src/DotNext/ValueTypeExtensions.cs
@@ -769,7 +769,7 @@ namespace DotNext
         public static double Normalize(this ulong value)
         {
             const ulong fraction = ulong.MaxValue >> (64 - 53);
-            const double exponent = (double)(1UL << 53);
+            const double exponent = 1L << 53;
             return (fraction & value) / exponent;
         }
 
@@ -790,7 +790,7 @@ namespace DotNext
         public static float Normalize(this uint value)
         {
             const uint fraction = uint.MaxValue >> (32 - 24);
-            const float exponent = (float)(1U << 24);
+            const float exponent = 1 << 24;
             return (fraction & value) / exponent;
         }
